Added FOIA_2017-004225-Factbase-Extended-Data.xlsx
Adds two tabs to sheet. One with the specifications and range of each aircraft. The other extends the information provided: clean columns for IATA and ICAO codes. Full airport name with lat/lon, city, state, and country, timezone information, local time, duration of the trip, and notes indicating a match to planespotter database and known trips.